+++ src/config/config.js
@@ -47,7 +47,7 @@ module.exports = {
         redis: {
             host: 'localhost',
             port: 6379,
-            prefix: 'templatemanager'
+            namespace: 'templatemanager'
         }
     },
     test: {