+++ src/config/config.pro.js
@@ -40,7 +40,7 @@ module.exports = {
         redis: {
             host: 'localhost',
             port: 6379,
-            prefix: 'templatemanager'
+            namespace: 'templatemanager'
         }
     },
     test: {
@@ -84,7 +84,7 @@ module.exports = {
         redis: {
             host: 'localhost',
             port: 6379,
-            prefix: 'zoodemo'
+            namespace: 'zoodemo'
         }
     }
 };